Refaktoroi: Vertailukoodia testeistä uudelleenkäytettäväksi;
Samoja tietorakenteita käytetään myös OAuth2-jaoissa, joiden testikoodissa voi käyttää samoja
vertailumetodeja.
